+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -43,6 +43,7 @@ interface AuthConfig extends ApiClientConfigParams {
     clientSecret?: string
     silenceWarnings?: boolean
     logger: Logger
+    defaultDnt?: boolean
     expirationTimeGuestRefreshToken?: number
     expirationTimeRegisteredRefreshToken?: number
 }
@@ -185,6 +186,7 @@ class Auth {
     private clientSecret: string
     private silenceWarnings: boolean
     private logger: Logger
+    private defaultDnt: boolean | undefined
     private expirationTimeGuestRefreshToken: number | undefined
     private expirationTimeRegisteredRefreshToken: number | undefined
 
@@ -235,6 +237,8 @@ class Auth {
         this.OCAPISessionsURL = config.OCAPISessionsURL || ''
 
         this.logger = config.logger
+
+        this.defaultDnt = config.defaultDnt
 
         if (
             config.expirationTimeGuestRefreshToken &&
@@ -493,7 +497,10 @@ class Auth {
                     () =>
                         helpers.refreshAccessToken(
                             this.client,
-                            {refreshToken},
+                            {
+                                refreshToken,
+                                ...(this.defaultDnt !== undefined && {dnt: this.defaultDnt})
+                            },
                             {
                                 clientSecret: this.clientSecret
                             }
@@ -543,12 +550,19 @@ class Auth {
         const isGuest = true
         const guestPrivateArgs = [
             this.client,
-            {...(usid && {usid})},
+            {
+                ...(this.defaultDnt !== undefined && {dnt: this.defaultDnt}),
+                ...(usid && {usid})
+            },
             {clientSecret: this.clientSecret}
         ] as const
         const guestPublicArgs = [
             this.client,
-            {redirectURI: this.redirectURI, ...(usid && {usid})}
+            {
+                redirectURI: this.redirectURI,
+                ...(this.defaultDnt !== undefined && {dnt: this.defaultDnt}),
+                ...(usid && {usid})
+            }
         ] as const
         const callback = this.clientSecret
             ? () => helpers.loginGuestUserPrivate(...guestPrivateArgs)
@@ -606,6 +620,7 @@ class Auth {
             },
             {
                 redirectURI,
+                ...(this.defaultDnt !== undefined && {dnt: this.defaultDnt}),
                 ...(usid && {usid})
             }
         )