+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v3.1.0-dev (Aug 08, 2024)
 
+- Add `defaultDnt` to support setting the dnt flag for SLAS. [#1979](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1979)
+
+
 ## v3.0.0 (Aug 07, 2024)
 - Add `meta.displayName` to queries. It can be used to identify queries in performance metrics or logs. [#1895](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1895)
 - Upgrade to commerce-sdk-isomorphic v3.0.0 [#1914](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1914)